fix: smooth line and graph output
Problem:

An extra line connected the last point to the first point in the graph
output.

Cause:

`vertSize` in `glDrawArrays(GL_LINE_STRIP_ADJACENCY, 0, vertSize);`, which
is the size of the vector `wave_temp` containing the vertices, whereas the
last parameter  of `glDrawArrays` was supposed to be the number of vertices.

Solution:

Changing `glDrawArrays(GL_LINE_STRIP_ADJACENCY, 0, vertSize);` to
`glDrawArrays(GL_LINE_STRIP_ADJACENCY, 0, vertSize / 2);`

Dividing vertSize by 2 gives the number of vertices.