fix(calculator): infinity bug
Note:
The calculator doesn't calculate inf for trigonometric waves like tan(x)
because the step size doesn't exactly reach a multiple of pi. So infinity
in the following contexts mean values that lie out of the scale of the
graph.

Problem:
When the expressions output have positive infinity and negative infinity
consecutively or a negative infinity and a positive infinity consecutively,
a line would be drawn from the positive to the negative
infinity.

Solution:
Introduce nan at either the positive or negative infinity for such infinity
pairs but not at both.